refactor: modularize the app store behind one persistence shell
