+++ src/store/initialState.ts
@@ -1,0 +1,119 @@
+
+import type { AppState, TrendingTimeRange } from '../types';
+import { defaultHeaderMenuConfig, defaultReleaseSourceSettings, defaultSubscriptionChannels } from '../types';
+import { DEFAULT_THEME_PRESET_ID } from '../constants/themePresets';
+import { readSessionBackendSecret } from './persistence/authStorage';
+import {
+  defaultDiscoveryChannels,
+  defaultMcpConfig,
+  defaultPresetFilters,
+  defaultVectorSearchConfig,
+  defaultVectorSearchStatus,
+  initialGistSearchFilters,
+  initialSearchFilters,
+} from './schema';
+
+export const createInitialState = (): AppState => ({
+      // Initial state
+      user: null,
+      githubToken: null,
+      isAuthenticated: false,
+      repositories: [],
+      gists: [],
+      starredGists: [],
+      gistSearchFilters: initialGistSearchFilters,
+      gistSearchResults: [],
+      selectedGistCategory: 'all',
+      analyzingGistIds: new Set<string>(),
+      isLoading: false,
+      isSyncingStars: false,
+      lastSync: null,
+      syncMode: 'stars',
+      syncModeConfigured: false,
+      listsPush: { isRunning: false, total: 0, done: 0, currentLabel: null, message: null, error: null },
+      categoryListIdMap: {},
+      analyzingRepositoryIds: new Set<number>(),
+      aiConfigs: [],
+      activeAIConfig: null,
+      embeddingConfigs: [],
+      activeEmbeddingConfig: null,
+      vectorSearchConfig: { ...defaultVectorSearchConfig },
+      vectorSearchStatus: { ...defaultVectorSearchStatus },
+      vectorIndexingState: { isIndexing: false, phase: null, phaseDone: 0, phaseTotal: 0, result: null },
+      mcpConfig: { ...defaultMcpConfig },
+      similarView: null,
+      webdavConfigs: [],
+      activeWebDAVConfig: null,
+      lastBackup: null,
+      searchFilters: initialSearchFilters,
+      searchResults: [],
+      releases: [],
+      releaseSubscriptions: new Set<number>(),
+      releaseSourceSettings: defaultReleaseSourceSettings,
+      readReleases: new Set<number>(),
+      customCategories: [],
+      hiddenDefaultCategoryIds: [],
+      defaultCategoryOverrides: {},
+      categoryOrder: [],
+      collapsedSidebarCategoryCount: 20,
+      categoryMatchMode: 'effective',
+      assetFilters: defaultPresetFilters,
+      theme: 'dark',
+      themePreset: DEFAULT_THEME_PRESET_ID,
+      hasHydrated: false,
+      currentView: 'repositories',
+      selectedCategory: 'all',
+      language: 'zh',
+      translationEngine: 'microsoft',
+      updateNotification: null,
+      analysisProgress: { current: 0, total: 0 },
+      backendApiSecret: readSessionBackendSecret(),
+      proxyConfig: { enabled: false, type: 'http', host: '', port: 7890 },
+      rpcDownloadConfig: { enabled: false, host: '', port: 6800 },
+      isSidebarCollapsed: false,
+      readmeModalOpen: false,
+      repositoryViewMode: 'grid',
+      releaseViewMode: 'timeline',
+      releaseShowMode: 'all',
+      releaseLatestMode: 'all',
+      releaseSelectedFilters: [],
+      releaseSearchQuery: '',
+      releaseExpandedRepositories: new Set<number>(),
+      releaseIsRefreshing: false,
+      includePreRelease: true,
+      includeKeysInBackup: false,
+
+      forks: [],
+      readForks: new Set<number>(),
+      forkViewMode: 'timeline',
+      forkSelectedFilters: [],
+      forkSearchQuery: '',
+      forkExpandedRepositories: new Set<number>(),
+      forkIsRefreshing: false,
+
+      discoveryChannels: defaultDiscoveryChannels,
+      discoveryRepos: { 'trending': [], 'hot-release': [], 'most-popular': [], 'topic': [], 'search': [] },
+      discoveryLastRefresh: { 'trending': null, 'hot-release': null, 'most-popular': null, 'topic': null, 'search': null },
+      discoveryIsLoading: { 'trending': false, 'hot-release': false, 'most-popular': false, 'topic': false, 'search': false },
+      discoveryIsLoadingMore: { 'trending': false, 'hot-release': false, 'most-popular': false, 'topic': false, 'search': false },
+      discoveryLoadMoreError: { 'trending': null, 'hot-release': null, 'most-popular': null, 'topic': null, 'search': null },
+      selectedDiscoveryChannel: 'trending',
+      discoveryPlatform: 'All',
+      discoveryLanguage: 'All',
+      discoverySortBy: 'BestMatch',
+      discoverySortOrder: 'Descending',
+      discoverySearchQuery: '',
+      discoverySelectedTopic: null,
+      discoveryHasMore: { 'trending': false, 'hot-release': false, 'most-popular': false, 'topic': false, 'search': false },
+      discoveryNextPage: { 'trending': 1, 'hot-release': 1, 'most-popular': 1, 'topic': 1, 'search': 1 },
+      discoveryTotalCount: { 'trending': 0, 'hot-release': 0, 'most-popular': 0, 'topic': 0, 'search': 0 },
+      discoveryScrollPositions: { 'trending': 0, 'hot-release': 0, 'most-popular': 0, 'topic': 0, 'search': 0 },
+  trendingTimeRange: 'weekly' as TrendingTimeRange,
+
+      // Subscription
+      subscriptionRepos: { 'most-stars': [], 'most-forks': [], 'most-dev': [], 'trending': [] },
+      subscriptionLastRefresh: { 'most-stars': null, 'most-forks': null, 'most-dev': null, 'trending': null },
+      subscriptionIsLoading: { 'most-stars': false, 'most-forks': false, 'most-dev': false, 'trending': false },
+      subscriptionChannels: defaultSubscriptionChannels,
+      headerMenuConfig: defaultHeaderMenuConfig,
+});
